KYLO-289 Show an error message if deleting a data source that is used by a feed.

+++ ui/ui-app/src/main/resources/static/js/feed-mgr/datasources/DatasourcesDetailsController.js
@@ -79,24 +79,35 @@ define(["angular", "feed-mgr/datasources/module-name"], function (angular, modul
          * Deletes the current data source.
          */
         self.onDelete = function () {
-            DatasourcesService.deleteById(self.model.id)
-                .then(function () {
-                    $mdToast.show(
-                        $mdToast.simple()
-                            .textContent("Successfully deleted the data source " + self.model.name + ".")
-                            .hideDelay(3000)
-                    );
-                    StateService.FeedManager().Datasource().navigateToDatasources();
-                }, function (err) {
-                    $mdDialog.show(
-                        $mdDialog.alert()
-                            .clickOutsideToClose(true)
-                            .title("Delete Failed")
-                            .textContent("The data source '" + self.model.name + "' could not be deleted. " + err.data.message)
-                            .ariaLabel("Failed to delete data source")
-                            .ok("Got it!")
-                    );
-                });
+            if (!angular.isArray(self.model.sourceForFeeds) || self.model.sourceForFeeds.length === 0) {
+                DatasourcesService.deleteById(self.model.id)
+                    .then(function () {
+                        $mdToast.show(
+                            $mdToast.simple()
+                                .textContent("Successfully deleted the data source " + self.model.name + ".")
+                                .hideDelay(3000)
+                        );
+                        StateService.FeedManager().Datasource().navigateToDatasources();
+                    }, function (err) {
+                        $mdDialog.show(
+                            $mdDialog.alert()
+                                .clickOutsideToClose(true)
+                                .title("Delete Failed")
+                                .textContent("The data source '" + self.model.name + "' could not be deleted." + err.data.message)
+                                .ariaLabel("Failed to delete data source")
+                                .ok("Got it!")
+                        );
+                    });
+            } else {
+                $mdDialog.show(
+                    $mdDialog.alert()
+                        .clickOutsideToClose(true)
+                        .title("Delete Failed")
+                        .textContent("This data source is currently being used by " + self.model.sourceForFeeds.length + " feed(s).")
+                        .ariaLabel("Failed to delete data source")
+                        .ok("Got it!")
+                );
+            }
         };
 
         /**
